refactor: enforce Steiger import locality
