+++ src/app/auth/index.tsx
@@ -1,8 +1,9 @@
 import React from "react";
 
-import { startAuthSession } from "@/app/auth/lifecycle";
 import { useAuthSession } from "@/entities/auth";
 import { RouteFeedback } from "@/shared/ui/route-feedback";
+
+import { startAuthSession } from "./lifecycle";
 
 export interface AuthProviderProps {
   children: React.ReactNode;
