+++ src/shared/ui/content/RemovableTag.tsx
@@ -6,7 +6,7 @@
 
 import type * as React from "react";
 
-import { TagMarker, tagClassName } from "@/shared/ui/content/tagStyles";
+import { TagMarker, tagClassName } from "./tagStyles";
 
 export interface RemovableTagProps {
   className?: string;
