+++ src/shared/ui/content/TagList.stories.tsx
@@ -6,9 +6,10 @@
 
 import type { Meta, StoryObj } from "@storybook/react";
 
-import { Tag } from "@/shared/ui/forms";
-import { TagList as Template } from "@/shared/ui/content/TagList";
 import * as fixture from "@/storybook/fixture";
+
+import { Tag } from "../forms";
+import { TagList as Template } from "./TagList";
 
 const meta = {
   title: "Shared/Content/TagList",
