+++ src/shared/ui/forms/Slider.stories.tsx
@@ -8,7 +8,7 @@ import type { Meta, StoryObj } from "@storybook/react";
 import * as React from "react";
 import { expect, fireEvent, fn } from "storybook/test";
 
-import { Slider as Template } from "@/shared/ui/forms/Slider";
+import { Slider as Template } from "./Slider";
 
 const meta = {
   title: "Shared/Forms/Slider",
